requirements.txt
flask==2.0.1
werkzeug==2.0.3
flask-cors==3.0.10
python-dotenv==1.0.0
requests==2.28.2
gunicorn==21.2.0
numpy>=1.20.0
pinecone-client>=3.0.0
langchain>=0.2.0
langchain-groq>=0.1.9
langchain-community>=0.2.0
langchain-core>=0.2.0
langchain-pinecone>=0.1.1
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sentence-transformers
huggingface_hub